+++ admin/modul/fasilitas/proses.php
@@ -12,16 +12,15 @@ if($aksi == 'tambah'){
     header("location:index.php");
 
 } elseif($aksi == 'edit'){
-    $id   = $_POST['id'];
+    $id = $_POST['id'];
     $nama = mysqli_real_escape_string($conn, $_POST['nama_fasilitas']);
     $desk = mysqli_real_escape_string($conn, $_POST['deskripsi']);
     $gambar = $_FILES['gambar']['name'];
 
     if($gambar != ""){
-        // Padam gambar lama
         $old = mysqli_fetch_array(mysqli_query($conn, "SELECT gambar FROM fasilitas WHERE id='$id'"));
-        unlink("../../../assets/img/".$old['gambar']);
-        // Upload baru
+        if(file_exists("../../../assets/img/".$old['gambar'])) unlink("../../../assets/img/".$old['gambar']);
+        
         move_uploaded_file($_FILES['gambar']['tmp_name'], "../../../assets/img/".$gambar);
         $sql = "UPDATE fasilitas SET nama_fasilitas='$nama', deskripsi='$desk', gambar='$gambar' WHERE id='$id'";
     } else {
@@ -33,7 +32,8 @@ if($aksi == 'tambah'){
 } elseif($aksi == 'hapus'){
     $id = $_GET['id'];
     $old = mysqli_fetch_array(mysqli_query($conn, "SELECT gambar FROM fasilitas WHERE id='$id'"));
-    unlink("../../../assets/img/".$old['gambar']);
+    if(file_exists("../../../assets/img/".$old['gambar'])) unlink("../../../assets/img/".$old['gambar']);
+    
     mysqli_query($conn, "DELETE FROM fasilitas WHERE id='$id'");
     header("location:index.php");
 }
